--- a/lib/tools.js
+++ b/lib/tools.js
@@ -31,6 +31,8 @@ function applyOptions (opts) {
   this.cache = opts.cache
   this.formatOpts = opts.formatOpts
   this.onTerminated = opts.onTerminated
+  this.levelMap = opts.levelMap
+  this.children = opts.children
 
   if (opts.level && opts.levelVal) {
     var levelIsStandard = levels.isStandardLevel(opts.level)
@@ -118,11 +120,24 @@ function genLog (z) {
   }
 }
 
+function byPrecision (b, a) {
+  var aix = a.indexOf('*')
+  var bix = b.indexOf('*')
+  if (aix === -1 && bix === -1) return 0
+  if (~aix && ~bix) {
+    if (aix > bix) return -1
+    if (aix < bix) return 1
+    return a.length < b.length ? 1 : (a.length === b.length ? 0 : -1)
+  }
+  return ~bix ? -1 : 1
+}
+
 module.exports = {
   noop: noop,
   copy: copy,
   applyOptions: applyOptions,
   defineLevelsProperty: defineLevelsProperty,
   streamIsBlockable: streamIsBlockable,
+  byPrecision: byPrecision,
   genLog: genLog
 }