--- a/lib/tools.js
+++ b/lib/tools.js
@@ -31,7 +31,7 @@ function applyOptions (opts) {
   this.cache = opts.cache
   this.formatOpts = opts.formatOpts
   this.onTerminated = opts.onTerminated
-  this.levelMap = opts.levelMap
+  this.levelRules = opts.levelRules
   this.children = opts.children
 
   if (opts.level && opts.levelVal) {
@@ -120,16 +120,21 @@ function genLog (z) {
   }
 }
 
-function byPrecision (b, a) {
+// ordering function for sorting of patterns: more specific patterns
+//  like 'myprefix:mylevel2:*' should be applied before 'myprefix*',
+// which itself should be applied before'*'
+function orderByPrecision (a, b) {
   var aix = a.indexOf('*')
   var bix = b.indexOf('*')
-  if (aix === -1 && bix === -1) return 0
   if (~aix && ~bix) {
-    if (aix > bix) return -1
-    if (aix < bix) return 1
-    return a.length < b.length ? 1 : (a.length === b.length ? 0 : -1)
+    return aix < bix ? 1 : aix === bix ? orderByPrecision(b.substr(aix+1), b.substr(bix+1)) : -1;
+  } else if (~aix) {
+    return 1;
+  } else if (~bix) {
+    return -1;
+  } else {
+    return a.length < b.length ? 1 : a.length === b.length ? 0 : -1;
   }
-  return ~bix ? -1 : 1
 }
 
 module.exports = {
@@ -138,6 +143,6 @@ module.exports = {
   applyOptions: applyOptions,
   defineLevelsProperty: defineLevelsProperty,
   streamIsBlockable: streamIsBlockable,
-  byPrecision: byPrecision,
+  orderByPrecision: orderByPrecision,
   genLog: genLog
 }