--- a/lib/tools.js
+++ b/lib/tools.js
@@ -127,13 +127,13 @@ function orderByPrecision (a, b) {
   var aix = a.indexOf('*')
   var bix = b.indexOf('*')
   if (~aix && ~bix) {
-    return aix < bix ? 1 : aix === bix ? orderByPrecision(b.substr(aix+1), b.substr(bix+1)) : -1;
+    return aix < bix ? 1 : aix === bix ? orderByPrecision(b.substr(aix + 1), b.substr(bix + 1)) : -1
   } else if (~aix) {
-    return 1;
+    return 1
   } else if (~bix) {
-    return -1;
+    return -1
   } else {
-    return a.length < b.length ? 1 : a.length === b.length ? 0 : -1;
+    return a.length < b.length ? 1 : a.length === b.length ? 0 : -1
   }
 }