--- a/test/helper.js
+++ b/test/helper.js
@@ -24,7 +24,8 @@ function check (t, chunk, level, msg) {
   })
 }
 
-function check2 (t, chunk, wanted) {
+// more generic version of check(), with 'wanted' an object
+function checkGen (t, chunk, wanted) {
   t.ok(new Date(chunk.time) <= new Date(), 'time is greater than Date.now()')
   delete chunk.time
   t.deepEqual(chunk, Object.assign({
@@ -36,4 +37,4 @@ function check2 (t, chunk, wanted) {
 
 module.exports.sink = sink
 module.exports.check = check
-module.exports.check2 = check2
+module.exports.checkGen = checkGen