As discussed in issue #206 an extension of the `logger.level` property to support fine-grained configuration of levels for a family of loggers.

Benchmark results are attached:
  [perf-org.txt](https://github.com/pinojs/pino/files/848062/perf-org.txt)
  [perf-new.txt](https://github.com/pinojs/pino/files/848061/perf-new.txt)