@jsumners 2017-03-16
I submitted my review incorrectly and can't seem to fix it. Consider them as "requested changes" not just comments.