@geonanorch 2017-03-16
Thanks for reviewing. I will try to add at least one test and address all comments &ndash;and requested-changes-as-comments