@geonanorch 2017-03-16
Apologies, pushed too early, unit test is not ready yet