@geonanorch 2017-03-16
@mcollina , I just committed 2 unit tests + a fix for dynamic changes.

Side-note: there is an issue (not related to my change) with the test `'level-change event'` in `levels.test.js`, it causes all subsequent tests to fail with:
```
    not ok 3 - test after end() was called # time=7.746ms
      ---
      stack: |
        EventEmitter.handle (test/levels.test.js:12:7)
        EventEmitter.setLevelVal (pino.js:125:12)
        EventEmitter._setLevel (pino.js:163:17)
        EventEmitter.applyOptions (lib/tools.js:43:8)
        EventEmitter.Pino (pino.js:104:22)
        pino (pino.js:91:16)
        Test.<anonymous> (test/levels.test.js:52:18)
        loop (node_modules/function-loop/index.js:35:15)
      at:
```
I did not try to fix it, do not understand enough about either `tap` or `pino`...