@jsumners 2017-03-16
@geonanorch that means you have introduced a bug. Look at the test that is failing. It is the one for setting the level by a number -- https://github.com/geonanorch/pino/blob/0224ab155cb9d55bf01b8b5c020b44baa2333f70/test/levels.test.js#L38