@geonanorch 2017-03-17
@jsumners the issue has been there beforehand &ndash;and CI is still happy. You can reproduce the problem _on your current master_ as follow:
  - edit `test/levels.test.js` to move the test `'level-change even'` anywhere else than at the bottom of the file
  - run full test suite through `npm` or even just `node_modules/.bin/tap test/levels.test.js`

Looks like that test is able to damage pino's or tap's state permanently, meaning there is a bug somewhere..Maybe I should have reported this through a separate channel...