@mcollina 2017-03-17
@geonanorch can you please open a separate issue on that broken test? If reordering makes your test pass, reordering is good.