@jsumners 2017-03-20
@geonanorch sorry about this, but I refactored the core code. You'll need to update your branch to incorporate the changes.