@geonanorch 2017-03-20
Ah... That's the problem with actively maintained projetcts... Unless you are just testing my motivation! ;-)
Anyway, changes reapplied and ready for your consideration.