@jsumners 2017-03-23
I'm about to publish a module that does this as a transport. I just have to work out a couple more tests.