@jsumners 2017-03-23
@geonanorch please check out https://github.com/pinojs/pino-filter . I haven't published it to npmjs.com yet. I want to make sure it works as you expect it to first.