@geonanorch 2017-03-23
@davidmarkclements benchmark results were included at the beginning of this issue. Since the same approach is used as for integer levels (logger functions set to noop in a _prior_ configuration call) a performance impact can only occur while creating children (2 internal object properties initialized with empty object/array, and one array iteration &ndash;empty array in your benchmark). And I nevertheless optimized child creation at @mcollina's request.

For this insignificant hit (feel free to verify by running the benchmarks) we get:
* improved coverage of the `debug` module to the point that it is no longer needed (I believe `pino` is only missing the timing info) 
* dynamic logging level configuration possible (unlike `debug`)

Together with the [Bunyan](https://github.com/trentm/node-bunyan) compatibility, I believe that it makes `pino` an excellent cover-all-options solution.

The branch is again in conflict. Let me know if you would like me to reapply the changes again &ndash;this time please only if you are ready to merge it right away, I already spent a lot of time on this as it is.

@jsumners  I had a quick look at the code of [pino-filter](https://github.com/pinojs/pino-filter). While functionally it almost certainly does what was discussed in issue #206  and in this thread, it involves testing the configuration for every logging call, which I reckon is far less efficient. What do the benchmark results look like?
Using signals for config changes is not compatible with window, the approach of [pino-arborsculpture](https://github.com/pinojs/pino-arborsculpture) (also by you!) is imho preferable.