@davidmarkclements 2017-03-23
Putting the transform in another process is more about keeping the event loop of the serving process free (and also keeps core as simple as possible, and also means less maintenance of browser pino to keep up)

On the other hand.. I do quite like the functionality as a core API ... 

I don't want to waste your time @geonanorch let's come to a definitive whether we want this in core 

I vote yes 

@jsumners  @mcollina  - yay or nay