@jsumners 2017-03-24
@geonanorch it doesn't really test the configuration for every logging call. It does determine if filters match, but it does so in a short-circuit fashion. As for reloading the config, I'm not a Windows guy so I do not consider it when doing things. But I have heard they have a Linux emulator in Win10 that Node stuff could be run under. Regardless, it's an initial implementation to prove the viability. If it ends up as the solution then it could certainly be improved (with your help?).

@davidmarkclements I leave that up to you and @mcollina. I have been quite clear that I think Pino has all of the features necessary to support this as a transport (and I have proved it so). That said, the benchmarks provided look very nice. Everything except for child creation is consistently below the master branch results. If nothing else, I'd like to see whatever changes caused those reductions to be included.

The questions that have me on the fence, so to speak, are:

1. Does this add too much complexity for maintenance?
2. Does this make things confusing for the user?

I truly do not know the answer to either of those. Thus my position.