@satazor 2017-03-26
I've been wanting a similar feature for quite some time.

What I want is to be able to propagate the loglevel of a parent to its children. This feature lets me do that, so +1 by me.