@mcollina 2017-03-27
If child creation gets on par with master, I'm +1 for something similar.

> Does this add too much complexity for maintenance?

A bit, but it seems a needed feature. If it does not carry a performance penalty, I think we can support it.

> Does this make things confusing for the user?

I do not like the fact that we accept a `pino.level = {}` property access. I think that should just be the actual level. Maybe we can have `setChildrenLevels()` or something similar.