@geonanorch 2017-03-28
Okay, I will reapply the changes to the current master (including latest requests for changes), can you confirm that there are no impending commits?
Re `setChildrenLevels()` no problems, note the  borderline case `setChildrenLevels('*': 'info'), which also sets the current logger's level (`setArborescenceLevels()` would be semantically a better match but too complicated I guess).