@geonanorch 2017-03-28
@mcollina I am not aware of any slowdown? I highlighted the non-impact 4 days ago &ndash;and proposed that you guys run the benchmarks yourselves for verification. Is there something specific you would like me to do?