@mcollina 2017-03-28
@geonanorch it seems @jsumners report a slowdown in child creation (conception benchmarks) 

> Everything except for child creation is consistently below the master branch results.

this PR:
```
benchPinoCreation*10000: 190.466ms
benchPinoExtremeCreation*10000: 114.567ms
```

master:
```
benchPinoCreation*10000: 183.904ms
benchPinoExtremeCreation*10000: 94.618ms
```