@davidmarkclements 2017-03-28
Just for context child creation is particular important for common scenarios like http - where pino-http creates a new child per request